# chore: upgrades golang from 1.11 to 1.17 based on faster performance