--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -1,0 +1,20 @@
+name: Go
+
+on:
+  - push
+  - pull_request
+
+jobs:
+
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Set up Go
+      uses: actions/setup-go@v2
+      with:
+        go-version: 1.17
+    - name: script/cibuild
+      run: |
+        script/cibuild