--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -7,14 +7,9 @@ on:
 jobs:
 
   build:
-    runs-on: ubuntu-latest
+    runs-on: golang:1.17
     steps:
     - uses: actions/checkout@v2
-
-    - name: Set up Go
-      uses: actions/setup-go@v2
-      with:
-        go-version: 1.11
     - name: script/cibuild
       run: |
         /bin/bash script/cibuild