--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -7,9 +7,16 @@ on:
 jobs:
 
   build:
-    runs-on: golang:1.17
+    runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
+
+    - name: Set up Go
+      uses: actions/setup-go@v2
+      with:
+        go-version: 1.17
+        
     - name: script/cibuild
       run: |
-        /bin/bash script/cibuild
+        ps -p $$
+        script/cibuild