--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -7,16 +7,15 @@ on:
 jobs:
 
   build:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-18.04
     steps:
     - uses: actions/checkout@v2
 
     - name: Set up Go
       uses: actions/setup-go@v2
       with:
-        go-version: 1.17
-        
+        go-version: 1.17.6
+
     - name: script/cibuild
       run: |
-        ps -p $$
         script/cibuild