--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -10,7 +10,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v2
-
+      with:
+        fetch-depth: 20
+        
     - name: Set up Go
       uses: actions/setup-go@v2
       with: