--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -11,8 +11,8 @@ jobs:
     steps:
     - uses: actions/checkout@v2
       with:
-        fetch-depth: 20
-        
+        fetch-depth: 0
+
     - name: Set up Go
       uses: actions/setup-go@v2
       with: