--- a/hesabu.go
+++ b/hesabu.go
@@ -124,7 +124,7 @@ func getInput(flag_arguments []string) ([]byte, error) {
 	}
 	var str []byte
 	fmt.Println("input mode", fi.Mode())
-	if fi.Mode()&os.ModeNamedPipe == 0 {
+	if len(flag_arguments) > 0 {
 		if len(flag_arguments) < 1 {
 			return nil, errors.New("No filename supplied")
 		}