--- a/script/build
+++ b/script/build
@@ -13,5 +13,5 @@ GOOS=linux GOARCH=amd64 go build -ldflags="-s -w" -o "${bindir}/hesabucli" hesab
 echo " -> New mac build at: ./bin/hesabucli-mac"
 echo " -> New build at: ./bin/hesabucli"
 
-./bin/hesabucli test/array.json
 echo '{"a": "1 + 2 + b", "b": "7"}' | ./bin/hesabucli
+cat test/array.json