--- a/script/build
+++ b/script/build
@@ -13,6 +13,4 @@ GOOS=linux GOARCH=amd64 go build -ldflags="-s -w" -o "${bindir}/hesabucli" hesab
 echo " -> New mac build at: ./bin/hesabucli-mac"
 echo " -> New build at: ./bin/hesabucli"
 
-echo '{"a": "1 + 2 + b", "b": "7"}' | ./bin/hesabucli
 ./bin/hesabucli -v
-./bin/hesabucli -d test/array.json