--- a/script/docker-build
+++ b/script/docker-build
@@ -1,2 +1,2 @@
 #!/bin/bash
-docker run --rm -v "$PWD":/usr/src/myapp -w /usr/src/myapp $1 golang:1.11 script/build
+docker run --rm -v "$PWD":/usr/src/myapp -w /usr/src/myapp $1 golang:1.17 script/build