--- a/test/array.json
+++ b/test/array.json
@@ -1,1 +1,4 @@
-{"without_array": "sum(1,2,3)",   "with_array": "sum(array(1,2,3))"}
+{
+    "without_array": "sum(1,2,3)",
+    "with_array": "sum(array(1,2,3))"
+}