--- a/test/example_runners.sh
+++ b/test/example_runners.sh
@@ -34,6 +34,7 @@ do
         echo "${name} \033[1;32mPASS\033[0m"
     else
         its_all_good=1
+        $cli $name
         echo "${name} \033[1;31mFAIL\033[0m"
     fi
 done