--- a/test/example_runners.sh
+++ b/test/example_runners.sh
@@ -18,7 +18,7 @@ fi
 
 for name in test/bad_*.json
 do
-    if $cli $name >/dev/null 2>&1
+    if $cli $name
     then
         its_all_good=1
         echo "${name} \033[1;31mFAIL\033[0m"
@@ -29,7 +29,7 @@ done
 
 for name in $(ls -1 test/*.json | grep -v 'bad_')
 do
-    if cat $name > $cli  >/dev/null 2>&1
+    if $cli $name
     then
         echo "${name} \033[1;32mPASS\033[0m"
     else
@@ -42,7 +42,7 @@ if [ "${its_all_good}" -gt "0" ]
 then
     echo "\n\033[1;31mSome examples are failing\033[0m"
 else
-    echo "\n\033[1;32mIt's all good\033[0m"
+    echo "\n\033[1;32mIts all good\033[0m"
 fi
 
 exit $its_all_good