--- a/test/example_runners.sh
+++ b/test/example_runners.sh
@@ -1,4 +1,4 @@
-#! /bin/sh
+#!/usr/bin/env bash
 #
 #
 set -ex