Performance tests revealed the following improvements for `real` with golang version 1.17: 

```golang 1.11
real	0m6,006s
user	0m7,057s
sys	0m0,365s

golang 1.17
real	0m4,386s
user	0m6,900s
sys	0m0,419s
```