@mestachs 2022-01-18
Thanks.
Strange I thought the build and tests were running.
Ok they used to be run by travis...  https://github.com/BLSQ/go-hesabu/blob/master/.travis.yml
Which probably doesn't have access anymore to our repos.
We can perhaps have a look to port this to github actions.